updated App.vue
In this commit update, the 'imageTag' component was removed from the project and a tag img was implemented in the 'app.vue'. Instead of using 'new Date', the 'intl' library was utilized for date formatting. Additionally, the API data was parsed and saved in a more convenient format for interpolation.